Размеры считать в том типе, в котором ими пользуются
Три умножения, результат которых уходит в size_t, но считается в
32 битах. Два безобидны — там константы; третье, размер буфера
миникарты, зависит от её стороны и при достаточно большой стороне
переполнится молча.

+++ app/src/main/cpp/src/ui/minimap.cpp
@@ -48,7 +48,7 @@ bool Minimap::init(vk::Context& ctx, u32 px) {
 
     dev_ = ctx.device();
     size_ = px;
-    pixels_.assign(size_ * size_ * 4, 0);
+    pixels_.assign((usize)size_ * size_ * 4, 0);
 
     if (!tex_.create(ctx.device(), ctx.physicalDevice(),
                      ctx.gfxQueue(), ctx.gfxFamily(),
